+++ Sources/TuistCore/Graph/Mappers/MapperEnvironmentKeys.swift
@@ -19,12 +19,8 @@ private struct TargetTestHashesKey: MapperEnvironmentKey {
     static var defaultValue: [AbsolutePath: [String: String]] = [:]
 }
 
-private struct CacheableTargetsKey: MapperEnvironmentKey {
-    static var defaultValue: [String] = []
-}
-
-private struct TargetHashesKey: MapperEnvironmentKey {
-    static var defaultValue: [CommandEventGraphTarget: String] = [:]
+private struct TargetTestCacheItemsKey: MapperEnvironmentKey {
+    static var defaultValue: [AbsolutePath: [String: CacheItem]] = [:]
 }
 
 extension MapperEnvironment {
@@ -34,9 +30,9 @@ extension MapperEnvironment {
         set { self[TargetTestHashesKey.self] = newValue }
     }
 
-    public var cacheableTargets: [String] {
-        get { self[CacheableTargetsKey.self] }
-        set { self[CacheableTargetsKey.self] = newValue }
+    public var targetTestCacheItems: [AbsolutePath: [String: CacheItem]] {
+        get { self[TargetTestCacheItemsKey.self] }
+        set { self[TargetTestCacheItemsKey.self] = newValue }
     }
 
     public var initialGraph: Graph? {
@@ -47,10 +43,5 @@ extension MapperEnvironment {
     public var initialGraphWithSources: Graph? {
         get { self[InitialGraphWithSourcesKey.self] }
         set { self[InitialGraphWithSourcesKey.self] = newValue }
-    }
-
-    public var targetHashes: [CommandEventGraphTarget: String] {
-        get { self[TargetHashesKey.self] }
-        set { self[TargetHashesKey.self] = newValue }
     }
 }