+++ Sources/TuistSupportTesting/Utils/MockEnvironment.swift
@@ -29,6 +29,10 @@ public final class MockEnvironment: Environmenting {
         directory.path.appending(components: ".cache")
     }
 
+    public var stateDirectory: AbsolutePath {
+        directory.path.appending(component: "state")
+    }
+
     public var queueDirectory: AbsolutePath {
         queueDirectoryStub ?? directory.path.appending(component: Constants.AsyncQueue.directoryName)
     }